+++ flix/flix.ino
@@ -13,7 +13,7 @@
 #define WIFI_ENABLED 0
 #define PIN_NEOPIXEL 38
 
-double t = NAN; // current step time, s
+float t = NAN; // current step time, s
 float dt; // time delta from previous step, s
 float controlRoll, controlPitch, controlYaw, controlThrottle; // pilot's inputs, range [-1, 1]
 float controlMode = NAN;
